+++ src/main/java/com/ixayda/iam/account/internal/AccountMfaStatusResponse.java
@@ -1,5 +1,5 @@
 package com.ixayda.iam.account.internal;
 
-record AccountMfaStatusResponse(boolean totpEnabled) {
+record AccountMfaStatusResponse(boolean totpEnabled, boolean recoveryCodesAvailable) {
 
 }